+++ app/src/main/java/com/example/wafarlytask/models/order_details_response_model/ProductInvetory.kt
@@ -3,10 +3,10 @@ package com.example.wafarlytask.models.order_details_response_model
 data class ProductInvetory(
     val id: Int,
     val inventoryId: Int,
-    val price: Int,
-    val priceAfterDiscount: Int,
+    val price: Double,
+    val priceAfterDiscount: Double,
     val productId: Int,
-    val quantity: Int,
+    val quantity: Double,
     val salableQuantity: Double,
     val status: Boolean
 )
