+++ app/src/main/java/com/example/wafarlytask/models/product_model/ProductInvetory.kt
@@ -3,8 +3,8 @@ package com.example.wafarlytask.models.product_model
 data class ProductInvetory(
     val id: Int,
     val inventoryId: Int,
-    val price: Int,
-    val priceAfterDiscount: Int,
+    val price: Double,
+    val priceAfterDiscount: Double,
     val productId: Int,
     val quantity: Double,
     val salableQuantity: Double,
